# docs/platform_token_api_contract_openapi_draft_v1_2026-03-29.yaml
openapi: 3.0.3
info:
  title: Platform Token Runtime API Contract Draft
  version: 1.0.0-draft
  description: |
    TOK-002/TOK-003/TOK-004 对应的 token 运行态接口草案。
    关键边界：
    1) 仅平台内部可调用 issue/revoke。
    2) 不暴露上游供应方凭证信息。
servers:
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
tags:
  - name: PlatformToken
paths:
  /api/v1/platform/tokens/issue:
    post:
      tags: [PlatformToken]
      summary: 签发平台短期 token
      operationId: issuePlatformToken
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueTokenRequest'
      responses:
        '201':
          description: 签发成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/platform/tokens/{tokenId}/refresh:
    post:
      tags: [PlatformToken]
      summary: 刷新 token 过期时间
      operationId: refreshPlatformToken
      parameters:
        - $ref: '#/components/parameters/TokenIdParam'
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/platform/tokens/{tokenId}/revoke:
    post:
      tags: [PlatformToken]
      summary: 吊销 token
      operationId: revokePlatformToken
      parameters:
        - $ref: '#/components/parameters/TokenIdParam'
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeTokenRequest'
      responses:
        '200':
          description: 吊销成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/platform/tokens/introspect:
    post:
      tags: [PlatformToken]
      summary: token 校验与解析
      operationId: introspectPlatformToken
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntrospectTokenRequest'
      responses:
        '200':
          description: 校验成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntrospectTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/BusinessError'
  /api/v1/platform/tokens/audit-events:
    get:
      tags: [PlatformToken]
      summary: 查询 token 审计事件
      operationId: listPlatformTokenAuditEvents
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - name: request_id
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: token_id
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: subject_id
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: event_name
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: result_code
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          $ref: '#/components/responses/BusinessError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    TokenIdParam:
      name: tokenId
      in: path
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessError:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    IssueTokenRequest:
      type: object
      required: [subject_id, role, ttl_seconds, scope]
      properties:
        subject_id:
          type: string
          minLength: 1
          maxLength: 128
        role:
          type: string
          enum: [owner, viewer, admin]
        ttl_seconds:
          type: integer
          minimum: 60
          maximum: 259200
        scope:
          type: array
          minItems: 1
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    IssueTokenResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [token_id, access_token, issued_at, expires_at, status]
          properties:
            token_id:
              type: string
            access_token:
              type: string
              writeOnly: true
            issued_at:
              type: string
              format: date-time
            expires_at:
              type: string
              format: date-time
            status:
              type: string
              enum: [active]
    RefreshTokenRequest:
      type: object
      required: [ttl_seconds]
      properties:
        ttl_seconds:
          type: integer
          minimum: 60
          maximum: 259200
    RefreshTokenResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [token_id, previous_expires_at, expires_at, status]
          properties:
            token_id:
              type: string
            previous_expires_at:
              type: string
              format: date-time
            expires_at:
              type: string
              format: date-time
            status:
              type: string
              enum: [active]
    RevokeTokenRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 256
    RevokeTokenResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [token_id, status, revoked_at]
          properties:
            token_id:
              type: string
            status:
              type: string
              enum: [revoked]
            revoked_at:
              type: string
              format: date-time
    IntrospectTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          minLength: 8
    IntrospectTokenResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [token_id, subject_id, role, status, scope, issued_at, expires_at]
          properties:
            token_id:
              type: string
            subject_id:
              type: string
            role:
              type: string
              enum: [owner, viewer, admin]
            status:
              type: string
              enum: [active, revoked, expired]
            scope:
              type: array
              items:
                type: string
            issued_at:
              type: string
              format: date-time
            expires_at:
              type: string
              format: date-time
    AuditEventsResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          type: string
        data:
          type: object
          required: [total, items]
          properties:
            total:
              type: integer
              minimum: 0
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditEventItem'
    AuditEventItem:
      type: object
      required: [event_id, event_name, request_id, route, result_code, created_at]
      properties:
        event_id:
          type: string
        event_name:
          type: string
        request_id:
          type: string
        token_id:
          type: string
        subject_id:
          type: string
        route:
          type: string
        result_code:
          type: string
        client_ip:
          type: string
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true